+++ src/components/block/api.ts
@@ -116,6 +116,14 @@ function BlockAPI(
     validate(data: BlockToolData): Promise<boolean> {
       return block.validate(data);
     },
+
+    /**
+     * Allows to say Editor that Block was changed. Used to manually trigger Editor's 'onChange' callback
+     * Can be useful for block changes invisible for editor core.
+     */
+    dispatchChange(): void {
+      block.dispatchChange();
+    },
   };
 
   Object.setPrototypeOf(this, blockAPI);