+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 2.23.0
 
 - `Improvement` — The `onChange` callback now accepts two arguments: EditorJS API and the CustomEvent with `type` and `detail` allowing to determine what happened with a Block
+- `New` *Block API* — The new `dispatchChange()` method allows to manually trigger the 'onChange' callback. Useful when Tool made a state mutation that is invisible for editor core.
 
 ### 2.22.3